Corretti alcuni bag dei calcolatori

+++ ProgettoGIA/Model/Calcolatori/CalcolatoreDYN.cs
@@ -61,7 +61,7 @@
                 temp = temp + p.ValutazioneTecnica_acquaticità + p.ValutazioneTecnica_assetto + p.ValutazioneTecnica_avanzamento + p.ValutazioneTecnica_virata;
 
                 //Penalità percentuali
-                temp = temp * (1 - (p.Penalità / 100)));
+                temp = temp * (1 - ((float)p.Penalità / 100));
 
                 p.Punteggio = temp;
                 p.IsCompletata = true;
